Make sure users cannot delete themselves. This should make sure there is at least one user in the system.

+++ app/Http/Controllers/UsersController.php
@@ -86,6 +86,10 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if ($user->id === auth()->user()->id) {
+            return response(['message' => 'You may not delete yourself.'], 422);
+        }
+
         $user->delete();
 
         return response([], 204);
